backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
# pydantic[email]==2.5.0
python-dotenv==1.0.0
# pydantic-settings==2.1.0
redis==5.0.1
aio-pika==9.3.1
celery==5.3.4
websockets==12.0
aiohttp==3.9.1